Update amount type to Double and balance to Numeric

+++ migrations/scripts/actual_collectable_prices.sql
@@ -1,6 +1,6 @@
 CREATE TABLE categories(
     price_id SERIAL PRIMARY KEY,
     price INT,
-    collectable_id INT,
+    collectable_id NUMERIC(15, 2),
     FOREIGN KEY (collectable_id) REFERENCES collectibles(collectible_id)
 );
